HBX-1584: Move class 'org.hibernate.cfg.MetaDataDialectFactory' to 'org.hibernate.tool.api.reveng.MetaDataDialectFactory'
