+++ test/nodb/src/test/java/org/hibernate/tool/cfg/MetaDataDialectFactoryTest/TestCase.java
@@ -3,7 +3,6 @@ package org.hibernate.tool.cfg.MetaDataDialectFactoryTest;
 import java.util.Properties;
 
 import org.hibernate.cfg.JDBCBinderException;
-import org.hibernate.cfg.MetaDataDialectFactory;
 import org.hibernate.cfg.reveng.dialect.H2MetaDataDialect;
 import org.hibernate.cfg.reveng.dialect.HSQLMetaDataDialect;
 import org.hibernate.cfg.reveng.dialect.JDBCMetaDataDialect;
@@ -17,6 +16,7 @@ import org.hibernate.dialect.MySQLDialect;
 import org.hibernate.dialect.Oracle10gDialect;
 import org.hibernate.dialect.Oracle8iDialect;
 import org.hibernate.dialect.Oracle9iDialect;
+import org.hibernate.tool.api.reveng.MetaDataDialectFactory;
 import org.junit.Assert;
 import org.junit.Test;
